+++ CHANGELOG.md
@@ -1,3 +1,14 @@
 # Changelog
 
 All notable changes to `php-slack` will be documented in this file.
+
+1.0.0 - 2025-02-08
+## What's changed
+* Document how to use the BlockKit facilities
+* Add test coverage for sending messages
+* Remove unfinished features
+
+---
+
+0.0.7 - 2025-02-01
+* Initial release